Create invalid user model
