+++ src/test/java/com/auto/test/sanity/LoginPageNegativeTest.java
@@ -1,6 +1,6 @@
 package com.auto.test.sanity;
 
-import com.auto.model.User;
+import com.auto.model.InvalidUser;
 import com.auto.model.UserModel;
 import com.auto.page.ILoginPage;
 import com.auto.page.PageFactory;
@@ -8,7 +8,6 @@ import com.auto.test.BrowserTestBase;
 import com.auto.utils.Assertion;
 import com.auto.utils.DriverUtils;
 import com.auto.utils.MessageLoader;
-import com.logigear.statics.Selaium;
 import org.testng.annotations.*;
 
 public class LoginPageNegativeTest extends BrowserTestBase {
@@ -18,7 +17,7 @@ public class LoginPageNegativeTest extends BrowserTestBase {
 
     @Test(description = "Fail to login specific repository successfully via Dashboard login page with incorrect credentials")
     public void DA_LOGIN_TC002() {
-        invalidUser = User.instance().getInvalidUser(0);
+        invalidUser = InvalidUser.instance().getInvalidUser(0);
         loginPage.enterUserAccount(invalidUser);
         loginPage.clickLoginButton();
         Assertion.assertEquals(loginPage.getAlertMessage(), MessageLoader.getMessage("invalid.username.password"), "Can not get alert message");
@@ -27,7 +26,7 @@ public class LoginPageNegativeTest extends BrowserTestBase {
 
     @Test(description = "Fail to log in successfully via Dashboard login page with correct username and incorrect password")
     public void DA_LOGIN_TC003() {
-        invalidUser = User.instance().getInvalidUser(1);
+        invalidUser = InvalidUser.instance().getInvalidUser(1);
         loginPage.enterUserAccount(invalidUser);
         loginPage.clickLoginButton();
         Assertion.assertEquals(loginPage.getAlertMessage(), MessageLoader.getMessage("invalid.username.password"), "Can not get alert message");
@@ -36,7 +35,7 @@ public class LoginPageNegativeTest extends BrowserTestBase {
 
     @Test(description = "Able to log in different repositories successfully after logging out current repository")
     public void DA_LOGIN_TC004() {
-        invalidUser = User.instance().getInvalidUser(1);
+        invalidUser = InvalidUser.instance().getInvalidUser(1);
         loginPage.enterUserAccount(invalidUser);
         loginPage.clickLoginButton();
         Assertion.assertEquals(loginPage.getAlertMessage(), MessageLoader.getMessage("invalid.username.password"), "Can not get alert message");
